remove service folder
